Simple Conjoint Analyses and Visualization
Source: R/package-cregg.R

Simple analyses of conjoint (factorial) experiments and visualization of results.

cj(data, formula, id = NULL, weights = NULL, estimate = c("amce", "freqs",
  "mm"), by = NULL, ...)

Arguments

data
A data frame containing variables specified in formula. All RHS variables should be factors; the base level for each will be used in estimation and for AMCEs the base level's AMCE will be zero.

formula
A formula specifying a model to be estimated. All variables should be factors.

id
An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.

weights
An (optional) RHS formula specifying a variable holding survey weights.

estimate
A character string specifying an estimate type. Current options are average marginal component effects (or AMCEs, “amce”, estimated via amce), display frequencies (“freq”, estimated via freqs), or marginal means (or AMMs, “mm”, estimated via mm). Additional options may be made available in the future.

by
A formula containing only RHS variables, specifying grouping factors over which to perform estimation.

…
Additional arguments to amce, freqs, or mm.

Details

The main function cj is a convenience function wrapper around the underlying estimation functions that provide for average marginal component effects (AMCEs), by default, via the amce function, marginal means (MMs) via the mm function, and display frequencies via freqs and props. Additional estimands may be supported in the future through their own functions and through the cj interface. Plotting is provided via ggplot2 for both types of estimates.

The only additional functionality provided by cj over the underlying functions is the by argument, which will perform operations on subsets of data, returning a single data frame. This can be useful, for example, for evaluating profile spillover effects and subgroup results.

See also

amce mm freqs plot.cj_amce

Examples

# NOT RUN {
# load data
requireNamespace("ggplot2")
data("hainmueller")

# calculate MMs
d1 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education +
         LanguageSkills + CountryOfOrigin + Job + JobExperience +
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID,
         estimate = "mm")

# plot MMs
plot(d1)

# MMs split by profile number
stacked <- do.call("rbind", lapply(1:3, function(x) {
    out <- mm(hainmueller[hainmueller$contest_no == x,], ChosenImmigrant ~ Gender +
              Education + LanguageSkills + CountryOfOrigin + Job + JobExperience +
              JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)
    out$contest <- x
    out
}))

## plot with grouping
plot(stacked, group = "contest", feature_headers = FALSE)

## plot with facetting
plot(stacked) + ggplot2::facet_wrap(~contest, nrow = 1L)

# estimate AMCEs
d2 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education +
         LanguageSkills + CountryOfOrigin + Job + JobExperience +
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)

# plot AMCEs
plot(d2)

# grouped operations
## examine profile ordering
x <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills,
        id = ~ CaseID, estimate = "mm", by = ~ contest_no)
plot(x, group = "contest_no", vline = 0.5)

## subgroup analysis
hainmueller$ethnosplit <- cut(hainmueller$ethnocentrism, 2)
x <- cj(na.omit(hainmueller), ChosenImmigrant ~ Gender + Education + LanguageSkills,
        id = ~ CaseID, estimate = "mm", by = ~ ethnosplit)
plot(x, group = "ethnosplit", vline = 0.5)
# }