Preference Heterogeneity Diagnostics

Tests for preference heterogeneity in conjoint experiments

amce_diffs(data, formula, by, id = NULL, weights = NULL,
  feature_order = NULL, feature_labels = NULL,
  level_order = c("ascending", "descending"), alpha = 0.05, ...)

cj_anova(data, formula, id = NULL, weights = NULL, by = NULL, ...)

Arguments

data
A data frame containing variables specified in formula. All RHS variables should be factors; the base level for each will be used in estimation and for AMCEs the base level's AMCE will be NA.

formula
A formula specifying a model to be estimated. All variables should be factors.

by
A formula containing only RHS variables, specifying grouping factors over which to perform estimation.

id
Ignored.

weights
An (optional) RHS formula specifying a variable holding survey weights.

…
Additional arguments to amce, freqs, or mm.

Value

amce_diffs returns a data frame similar to the one returned by cj, including a BY column for easy merging with results returned by that function. cj_anova returns an anova object.

Details

cj_anova takes a model formula (“reduced” model) and generates a “full” model with full saturated interactions between the variables specified in by and all RHS variables in formula, then computes an F-test comparing the two models, providing a test for whether preferences vary across levels of by. This is, in essence, a test of whether all such interaction coefficients are distinguishable from zero. Because the test depends on overall model fit, not the coefficient variances, clustering is irrelevant.

See also

Examples

data(hainmueller)
# Test for heterogeneity by profile order
cj_anova(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills, by = ~ contest_no)

#> Analysis of Deviance Table
#> 
#> Model 1: ChosenImmigrant ~ Gender + Education + LanguageSkills
#> Model 2: ChosenImmigrant ~ Gender + Education + LanguageSkills + contest_no + 
#>     Gender:contest_no + Education:contest_no + LanguageSkills:contest_no
#>   Resid. Df Resid. Dev Df Deviance      F Pr(>F)
#> 1     13949     3353.0                          
#> 2     13938     3349.6 11   3.3873 1.2814 0.2279

# Test for heterogeneity by CountryOfOrigin feature
cj_anova(hainmueller, ChosenImmigrant ~ Gender + Education, by = ~ CountryOfOrigin)

#> Analysis of Deviance Table
#> 
#> Model 1: ChosenImmigrant ~ Gender + Education
#> Model 2: ChosenImmigrant ~ Gender + Education + CountryOfOrigin + Gender:CountryOfOrigin + 
#>     Education:CountryOfOrigin
#>   Resid. Df Resid. Dev Df Deviance      F    Pr(>F)    
#> 1     13952     3407.8                                 
#> 2     13880     3374.5 72   33.286 1.9016 6.756e-06 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

# Differences in AMCEs by Gender feature (i.e., feature interactions)
amce_diffs(hainmueller, ChosenImmigrant ~ LanguageSkills + Education, ~ Gender, id = ~ CaseID)

#> Waiting for profiling to be done...

#>           BY         outcome                feature                    level
#> 1 Difference ChosenImmigrant Educational Attainment                4th Grade
#> 2 Difference ChosenImmigrant Educational Attainment                8th Grade
#> 3 Difference ChosenImmigrant Educational Attainment           College Degree
#> 4 Difference ChosenImmigrant Educational Attainment          Graduate Degree
#> 5 Difference ChosenImmigrant Educational Attainment              High School
#> 6 Difference ChosenImmigrant Educational Attainment         Two-Year College
#> 7 Difference ChosenImmigrant        Language Skills           Broken English
#> 8 Difference ChosenImmigrant        Language Skills Tried English but Unable
#> 9 Difference ChosenImmigrant        Language Skills         Used Interpreter
#>          Gender     estimate  std.error          t         p       lower
#> 1 Male - Female -0.005966999 0.03110785 -0.1918165 0.8478887 -0.06693727
#> 2 Male - Female -0.030625179 0.03118286 -0.9821157 0.3260599 -0.09174247
#> 3 Male - Female -0.014844848 0.03131950 -0.4739810 0.6355209 -0.07622994
#> 4 Male - Female -0.042143177 0.03110228 -1.3549867 0.1754437 -0.10310253
#> 5 Male - Female -0.013956325 0.03123056 -0.4468804 0.6549684 -0.07516710
#> 6 Male - Female -0.047183288 0.03115487 -1.5144753 0.1299280 -0.10824572
#> 7 Male - Female -0.029903918 0.02343386 -1.2760985 0.2019420 -0.07583345
#> 8 Male - Female  0.004543744 0.02333837  0.1946898 0.8456386 -0.04119863
#> 9 Male - Female -0.013293411 0.02342970 -0.5673744 0.5704690 -0.05921478
#>        upper
#> 1 0.05500327
#> 2 0.03049211
#> 3 0.04654025
#> 4 0.01881617
#> 5 0.04725445
#> 6 0.01387914
#> 7 0.01602561
#> 8 0.05028611
#> 9 0.03262796