Immigration Conjoint Experiment Dataset from Hainmueller et. al. (2014)
Source: R/hainmueller.R

A dataset containing the results of a conjoint survey of a representative sample of American adults who were asked to choose which hypothetical immigrants they think should be admitted into the United States. Each row corresponds to a single profile presented to the respondent.

hainmueller

Format

A data frame with 13960 observations on the following 16 variables.

CaseID

a numeric vector indicating the respondent to which the particular profile corresponds

contest_no

a numeric vector indicating the number of the task to which the profile corresponds

Education

a factor with levels “No formal”, “4th grade”, “8th grade”, “High school”, “Two-year college”, “college Degree”, “Graduate degree”

Gender

a factor with levels “Female”, “Male”

CountryOfOrigin

a factor with levels “India”, “Germany”, “France”, “Mexico”, “Philippines”, “Poland”, “China”, “Sudan”, “Somalia”, “Iraq”

ReasonForApplication

a factor with levels “Reunite with family”, “Seek better job”, “Escape persecution”

Job

a factor with levels “Janitor”, “Waiter”, “Child care provider”, “Gardener”, “Financial analyst”, “Construction worker”, “Teacher”, “Computer programmer”, “Nurse”, “Research scientist”, “Doctor”

JobExperience

a factor with levels “None”, “1-2 years”, “3-5 years”, “5+ years”

JobPlans

a factor with levels “Will look for work”, “Contract with employer”, “Interviews with employer”, “No plans to look for work”

PriorEntry

a factor with levels “Never”, “Once as tourist”, “Many times as tourist”, “Six months with family”, “Once w/o authorization”

LanguageSkills

a factor with levels “Fluent English”, “Broken English”, “Tried English but unable”, “Used interpreter”

ChosenImmigrant

a numeric vector denoting whether the immigrant profile was selected

ethnocentrism

a numeric vector

profile

a numeric vector giving the profile number

LangPos

a numeric vector

PriorPos

a numeric vector

Source

Hainmueller, J., Hopkins, D., and Yamamoto T. (2014) Causal Inference in Conjoint Analysis: Understanding Multi-Dimensional Choices via Stated Preference Experiments. Political Analysis 22(1):1-30

Note

This is a modified version of the dataset available from the cjoint package.

See also

amce mm mm

Examples

# NOT RUN {
data("hainmueller")
cj(hainmueller, ChosenImmigrant ~ Gender + Education, id = ~ CaseID)
# }