Plot AMCE estimates, MM descriptives, and frequency plots

ggplot2-based plotting of conjoint AMCEs estimates and MMs

# S3 method for cj_amce
plot(x, group = attr(x, "by"), feature_headers = TRUE,
  header_fmt = "(%s)", size = 1, xlab = "Estimated AMCE", ylab = "",
  legend_title = if (is.null(group)) "Feature" else group,
  legend_pos = "bottom", xlim = NULL, vline = 0, vline_color = "gray",
  theme = ggplot2::theme_bw(), ...)

Arguments

x
A data frame returned from cj or mm.

group
Optionally a character string specifying a grouping factor. This is useful when, for example, subgroup analyses or comparing AMCEs for different outcomes. An alternative is to use facet_wrap for faceted graphics.

feature_headers
A logical indicating whether to include headers for each feature to visually separate levels for each feature (beyond the color palette).

header_fmt
A character string specifying a fmt argument to sprintf, which will be used when generating the feature headers (if feature_headers = TRUE).

size
A numeric value specifying point size in geom_point.

xlab
A label for the x-axis

ylab
A label for the y-axis

legend_title
A character string specifying a label for the legend.

legend_pos
An argument forwarded to the legend.position argument in theme.

xlim
A two-element number vector specifying limits for the x-axis. If NULL, a default value is calculated from the data.

vline
Optionally, a numeric value specifying an x-intercept for a vertical line. This can be useful in distinguishing the midpoint of the estimates (e.g., a zero line for AMCEs).

vline_color
A character string specifying a color for the vline.

theme
A ggplot2 theme object

…
Ignored.

Value

A ggplot2 object

Details

Two plot methods are implemented: one for “amce” data frames of conjoint AMCE estimates (returned by amce), and another for “mm” data frames as returned by mm. Both functions are basically identical but vary slightly in their superficial features (axis labels, etc.).

See also

Examples

# NOT RUN {
# load data
data("hainmueller")

# calculate MMs
d1 <- mm(hainmueller, ChosenImmigrant ~ Gender + Education +
         LanguageSkills + CountryOfOrigin + Job + JobExperience +
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)

# plot MMs
## simple plot
plot(d1)
## plot with facetting by feature
plot(d1, feature_headers = FALSE) +
  ggplot2::facet_wrap(~feature, ncol = 1L,
                      scales = "free_y", strip.position = "right")

# MMs split by profile number
stacked <- cj(hainmueller, ChosenImmigrant ~ Gender +
              Education + LanguageSkills + CountryOfOrigin + Job + JobExperience +
              JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID,
              estimate = "mm", by = ~ contest_no)

## plot with grouping
plot(stacked, group = "contest_no", feature_headers = FALSE)

## plot with facetting
plot(stacked) + ggplot2::facet_wrap(~contest_no, nrow = 1L)

# estimate AMCEs
d2 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education +
         LanguageSkills + CountryOfOrigin + Job + JobExperience +
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)

# plot AMCEs
plot(d2)
# }