Tax Preference Conjoint Experiment Dataset from Ballard-Rosa et al. (2016)

A dataset containing the results of a fully randomized conjoint survey of a representative sample of 2000 American adults who were asked to choose between alternative tax rate policies. Variables taxrate1-taxrate6 refer to tax rates for different income brackets and taxrev refers to levels of total tax revenue.

taxes

Format

A data frame (with additional “cj_df” class) with 32000 observations on the following 13 variables. Each row corresponds to a single profile presented to a respondent.

chose_plan

A numeric vector denoting whether the immigrant profile was selected (=1) or not (=0).

taxrate1

An experimental factor with levels “<10k: 0%”, “<10k: 5%”, “<10k: 15%”, “<10k: 25%”.

taxrate2

An experimental factor with levels “10-35k: 5%”, “10-35k: 15%”, “10-35k: 25%”, “10-35k: 35%”.

taxrate3

An experimental factor with levels “35-85k: 5%”, “35-85k: 15%”, “35-85k: 25%”, “35-85k: 35%”.

taxrate4

An experimental factor with levels “85-175k: 5%”, “85-175k: 15%”, “85-175k: 25%”, “85-175k: 35%”.

taxrate5

An experimental factor with levels “175-375k: 5%”, “175-375k: 15%”, “175-375k: 25%”, “175-375k: 35%”, “175-375k: 45%”.

taxrate6

An experimental factor with levels “>375k: 5%”, “>375k: 15%”, “>375k: 25%”, “>375k: 35%”, “>375k: 45%”, “>375k: 55%”.

taxrev

An experimental factor with levels “<75%”, “75-95%”, “95-105%”, “105-125%”, “>125%”.

inequality_aversion

A covariate specifying whether respondent is inequality averse (=1) or not (=0).

taxes_harm_economy

A covariate specifying whether respondent believes taxes harm the economy (=1) or not (=0).

partyid

A factor specifying the respondent's party identification; one of “Independent”, “Democrat”, “Republican”.

ID

A numeric vector indicating the respondent to which the profile corresponds.

weight

A numeric vector containing survey weights.

Source

Ballard-Rosa, Cameron, Lucy Martin, and Kenneth Scheve. 2016. “The Structure of American Income Tax Policy Preferences.” The Journal of Politics 79(1): 1-16. http://doi.org/10.1086/687324

See also

Examples

# \donttest{
data("taxes")
f1 <- chose_plan ~ taxrate1 + taxrate2 + taxrate3 +
        taxrate4 + taxrate5 + taxrate6 + taxrev
cj(taxes, f1, id = ~ ID, weights = ~     weight)

#>       outcome statistic                        feature         level
#> 1  chose_plan      amce          Tax rate for <$10,000      <10k: 0%
#> 2  chose_plan      amce          Tax rate for <$10,000      <10k: 5%
#> 3  chose_plan      amce          Tax rate for <$10,000     <10k: 15%
#> 4  chose_plan      amce          Tax rate for <$10,000     <10k: 25%
#> 5  chose_plan      amce   Tax rate for $10,000-$35,000    10-35k: 5%
#> 6  chose_plan      amce   Tax rate for $10,000-$35,000   10-35k: 15%
#> 7  chose_plan      amce   Tax rate for $10,000-$35,000   10-35k: 25%
#> 8  chose_plan      amce   Tax rate for $10,000-$35,000   10-35k: 35%
#> 9  chose_plan      amce   Tax rate for $25,000-$85,000    35-85k: 5%
#> 10 chose_plan      amce   Tax rate for $25,000-$85,000   35-85k: 15%
#> 11 chose_plan      amce   Tax rate for $25,000-$85,000   35-85k: 25%
#> 12 chose_plan      amce   Tax rate for $25,000-$85,000   35-85k: 35%
#> 13 chose_plan      amce  Tax rate for $85,000-$175,000   85-175k: 5%
#> 14 chose_plan      amce  Tax rate for $85,000-$175,000  85-175k: 15%
#> 15 chose_plan      amce  Tax rate for $85,000-$175,000  85-175k: 25%
#> 16 chose_plan      amce  Tax rate for $85,000-$175,000  85-175k: 35%
#> 17 chose_plan      amce Tax rate for $175,000-$375,000  175-375k: 5%
#> 18 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 15%
#> 19 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 25%
#> 20 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 35%
#> 21 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 45%
#> 22 chose_plan      amce         Tax rate for >$375,000     >375k: 5%
#> 23 chose_plan      amce         Tax rate for >$375,000    >375k: 15%
#> 24 chose_plan      amce         Tax rate for >$375,000    >375k: 25%
#> 25 chose_plan      amce         Tax rate for >$375,000    >375k: 35%
#> 26 chose_plan      amce         Tax rate for >$375,000    >375k: 45%
#> 27 chose_plan      amce         Tax rate for >$375,000    >375k: 55%
#> 28 chose_plan      amce                    Tax revenue          <75%
#> 29 chose_plan      amce                    Tax revenue        75-95%
#> 30 chose_plan      amce                    Tax revenue       95-105%
#> 31 chose_plan      amce                    Tax revenue      105-125%
#> 32 chose_plan      amce                    Tax revenue         >125%
#>         estimate  std.error            z            p        lower        upper
#> 1   0.0000000000         NA           NA           NA           NA           NA
#> 2  -0.0213657909 0.01153916  -1.85159027 6.408468e-02 -0.043982123  0.001250541
#> 3  -0.1085089681 0.01345414  -8.06509750 7.317744e-16 -0.134878602 -0.082139334
#> 4  -0.2417205854 0.01756500 -13.76149232 4.344651e-43 -0.276147349 -0.207293822
#> 5   0.0000000000         NA           NA           NA           NA           NA
#> 6  -0.0216654811 0.01529549  -1.41646250 1.566401e-01 -0.051644081  0.008313119
#> 7  -0.1028048639 0.02277710  -4.51351823 6.376097e-06 -0.147447163 -0.058162565
#> 8  -0.2194266402 0.02894688  -7.58032053 3.447023e-14 -0.276161485 -0.162691795
#> 9   0.0000000000         NA           NA           NA           NA           NA
#> 10  0.0001812976 0.01151841   0.01573981 9.874420e-01 -0.022394376  0.022756971
#> 11 -0.0650669525 0.01446934  -4.49688323 6.895683e-06 -0.093426346 -0.036707558
#> 12 -0.1260348655 0.01762314  -7.15166921 8.572896e-13 -0.160575585 -0.091494146
#> 13  0.0000000000         NA           NA           NA           NA           NA
#> 14  0.0109438886 0.01084215   1.00938387 3.127906e-01 -0.010306330  0.032194107
#> 15 -0.0155419502 0.01134766  -1.36961717 1.708064e-01 -0.037782955  0.006699055
#> 16 -0.0126122794 0.01190376  -1.05952103 2.893626e-01 -0.035943212  0.010718653
#> 17  0.0000000000         NA           NA           NA           NA           NA
#> 18  0.0332708646 0.01162503   2.86200317 4.209727e-03  0.010486230  0.056055499
#> 19  0.0425251595 0.01305320   3.25783404 1.122660e-03  0.016941357  0.068108961
#> 20  0.0580764422 0.01212647   4.78922984 1.674226e-06  0.034309002  0.081843883
#> 21  0.0327895914 0.01402496   2.33794557 1.939007e-02  0.005301176  0.060278007
#> 22  0.0000000000         NA           NA           NA           NA           NA
#> 23  0.0558225384 0.01369058   4.07744003 4.553427e-05  0.028989485  0.082655591
#> 24  0.0896621795 0.01494775   5.99837302 1.993043e-09  0.060365128  0.118959231
#> 25  0.1139343820 0.01551745   7.34233988 2.098916e-13  0.083520742  0.144348022
#> 26  0.1099558977 0.01729319   6.35833663 2.039501e-10  0.076061876  0.143849919
#> 27  0.0838214633 0.01874157   4.47248772 7.731484e-06  0.047088652  0.120554275
#> 28  0.0000000000         NA           NA           NA           NA           NA
#> 29  0.0297942004 0.01755028   1.69764821 8.957418e-02 -0.004603715  0.064192116
#> 30  0.0692814207 0.02344570   2.95497381 3.126956e-03  0.023328699  0.115234143
#> 31  0.0911924955 0.03000830   3.03890950 2.374362e-03  0.032377316  0.150007675
#> 32  0.1354902736 0.03885990   3.48663428 4.891398e-04  0.059326261  0.211654286

# }