Tidy estimation of AMCEs

Estimate AMCEs for a conjoint analysis and return a tidy data frame of results

amce(data, formula, id = NULL, weights = NULL, feature_order = NULL,
  feature_labels = NULL, level_order = c("ascending", "descending"),
  alpha = 0.05, ...)

amce_by_reference(data, formula, variable, ...)

Arguments

data
A data frame containing variables specified in formula. All RHS variables should be factors; the base level for each will be used in estimation and its reported AMCE will be NA (for printing).

formula
A formula specifying an AMCE model to be estimated. All variables should be factors.

id
An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.

weights
An (optional) RHS formula specifying a variable holding survey weights.

feature_order
An (optional) character vector specifying the names of feature (RHS) variables in the order they should be encoded in the resulting data frame.

feature_labels
A named list of “fancy” feature labels to be used in output. By default, the function looks for a “label” attribute on each variable in formula and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables on the righthand side of formula and character string values; arguments passed here override variable attributes.

level_order
A character string specifying levels (within each feature) should be ordered increasing or decreasing in the final output. This is mostly only consequential for plotting via plot.cj_mm, etc.

alpha
A numeric value indicating the significance level at which to calculate confidence intervals for the MMs (by default 0.95, meaning 95-percent CIs are returned).

…
For amce: additional arguments to glm or svyglm, the latter being used if weights is non-NULL. For amce_by_reference: additional arguments passed to amce.

variable
An RHS formula containing a single factor variable from formula. This will be used by amce_by_reference to estimate AMCEs relative to each possible factor level as a reference category. If more than one RHS variables are specified, the first will be used.

Value

A data frame

Details

amce provides estimates of AMCEs (or rather, average marginal effects for each feature level). It does not calculate AMCEs for constrained conjoint designs. The function can also be used for balance testing by specifying a covariate rather outcome on the left-hand side of formula. See examples.

amce_by_reference provides a tool for quick sensitivity analysis. AMCEs are defined relative to an arbitrary reference category (i.e., feature level). This function will loop over all feature levels (for a specified feature) to show how interpretation will be affected by choice of reference category. The resulting data frame will be a stacked result from amce, containing an additional .reference column specifying which level of variable was used as the reference category.

Users may desire to specify a family argument via …, which should be a “family” object such as gaussian. Sensible alternatives are binomial (for binary outcomes) and quasibinomial (for weighted survey data).

See also

Examples

data(hainmueller)
# estimating AMCEs
amce(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills,
     id = ~ CaseID, feature_order = c("LanguageSkills", "Gender", "Education"))

#>            outcome                feature                    level    estimate
#> 1  ChosenImmigrant        Language Skills           Fluent English  0.00000000
#> 2  ChosenImmigrant        Language Skills           Broken English -0.05933832
#> 3  ChosenImmigrant        Language Skills Tried English but Unable -0.12746827
#> 4  ChosenImmigrant        Language Skills         Used Interpreter -0.16199396
#> 5  ChosenImmigrant                 Gender                   Female  0.00000000
#> 6  ChosenImmigrant                 Gender                     Male -0.02726620
#> 7  ChosenImmigrant Educational Attainment                No Formal  0.00000000
#> 8  ChosenImmigrant Educational Attainment                4th Grade  0.03139746
#> 9  ChosenImmigrant Educational Attainment                8th Grade  0.05694292
#> 10 ChosenImmigrant Educational Attainment              High School  0.11825338
#> 11 ChosenImmigrant Educational Attainment         Two-Year College  0.17622493
#> 12 ChosenImmigrant Educational Attainment           College Degree  0.19676219
#> 13 ChosenImmigrant Educational Attainment          Graduate Degree  0.19290717
#>      std.error          z            p        lower       upper
#> 1           NA         NA           NA           NA          NA
#> 2  0.011803743  -5.027076 4.980141e-07 -0.082473226 -0.03620341
#> 3  0.011855107 -10.752182 5.787623e-27 -0.150703848 -0.10423268
#> 4  0.011996863 -13.503027 1.500807e-41 -0.185507378 -0.13848054
#> 5           NA         NA           NA           NA          NA
#> 6  0.008355089  -3.263424 1.100748e-03 -0.043641869 -0.01089052
#> 7           NA         NA           NA           NA          NA
#> 8  0.015401407   2.038610 4.148899e-02  0.001211256  0.06158366
#> 9  0.015430629   3.690253 2.240315e-04  0.026699440  0.08718639
#> 10 0.015606298   7.577285 3.528603e-14  0.087665593  0.14884116
#> 11 0.015604750  11.293032 1.420507e-29  0.145640187  0.20680968
#> 12 0.015948173  12.337600 5.682440e-35  0.165504342  0.22802003
#> 13 0.015744658  12.252230 1.634391e-34  0.162048209  0.22376613

# NOT RUN {
# balance testing example
plot(amce(hainmueller[!is.na(hainmueller$ethnocentrism),
     ethnocentrism ~ Gender + Education + LanguageSkills, id = ~ CaseID))

# reference category sensitivity
x <- amce_by_reference(hainmueller, ChosenImmigrant ~ LanguageSkills + Education, 
       variable = ~ LanguageSkills, id = ~ CaseID)
plot(x, group = ".reference")
# }