Marginal Means

Calculate (descriptive) marginal means (MMs) from a conjoint design

mm(data, formula, id = NULL, weights = NULL, feature_order = NULL,
  feature_labels = NULL, level_order = c("ascending", "descending"),
  h0 = 0, alpha = 0.05, ...)

Arguments

data
A data frame containing variables specified in formula. All RHS variables should be factors.

formula
A formula specifying an outcome (LHS) and conjoint features (RHS) to describe. All RHS variables should be factors, with constraints specified with an asterisk (*) between features, as in amce.

id
An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.

weights
An (optional) RHS formula specifying a variable holding survey weights.

feature_order
An (optional) character vector specifying the names of feature (RHS) variables in the order they should be encoded in the resulting data frame.

feature_labels
A named list of “fancy” feature labels to be used in output. By default, the function looks for a “label” attribute on each variable in formula and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables on the righthand side of formula and character string values; arguments passed here override variable attributes.

level_order
A character string specifying levels (within each feature) should be ordered increasing or decreasing in the final output. This is mostly only consequential for plotting via plot.cj_mm, etc.

h0
A numeric value specifying a null hypothesis value to use when generating z-statistics and p-values.

alpha
A numeric value indicating the significance level at which to calculate confidence intervals for the MMs (by default 0.95, meaning 95-percent CIs are returned).

…
Ignored.

Details

mm provides descriptive representations of conjoint data as marginal means (MMs), which represent the mean outcome across all appearances of a particular conjoint feature level, averaging across all other features. In forced choice conjoint designs with two profiles per choice task, MMs by definition average 0.5 with values above 0.5 indicating features that increase profile favorability and values below 0.5 indicating features that decrease profile favorability. For continuous outcomes, MMs can take any value in the full range of the outcome.

But note that if feature levels can co-occur, such that both alternatives share a feature level, then the MMs on forced choice outcomes are bounded by the probability of co-occurrence (as a lower bound) and 1 minus that probability as an upper bound.

Plotting functionality is provided in plot.cj_mm.

See also

Examples

data(immigration)
# marginal means
mm(immigration, ChosenImmigrant ~ Gender + Education + LanguageSkills,
   id = ~ CaseID, h0 = 0.5)

#>            outcome statistic                feature                    level
#> 1  ChosenImmigrant        mm                 Gender                   Female
#> 2  ChosenImmigrant        mm                 Gender                     Male
#> 3  ChosenImmigrant        mm Educational Attainment                No Formal
#> 4  ChosenImmigrant        mm Educational Attainment                4th Grade
#> 5  ChosenImmigrant        mm Educational Attainment                8th Grade
#> 6  ChosenImmigrant        mm Educational Attainment              High School
#> 7  ChosenImmigrant        mm Educational Attainment         Two-Year College
#> 8  ChosenImmigrant        mm Educational Attainment           College Degree
#> 9  ChosenImmigrant        mm Educational Attainment          Graduate Degree
#> 10 ChosenImmigrant        mm        Language Skills           Fluent English
#> 11 ChosenImmigrant        mm        Language Skills           Broken English
#> 12 ChosenImmigrant        mm        Language Skills Tried English but Unable
#> 13 ChosenImmigrant        mm        Language Skills         Used Interpreter
#>     estimate   std.error           z            p     lower     upper
#> 1  0.5117812 0.005919838   1.9901230 4.657739e-02 0.5001785 0.5233839
#> 2  0.4877013 0.006048301  -2.0334110 4.201102e-02 0.4758469 0.4995558
#> 3  0.3895112 0.011003561 -10.0411853 1.004597e-23 0.3679446 0.4110778
#> 4  0.4222002 0.010994931  -7.0759699 1.484071e-12 0.4006505 0.4437499
#> 5  0.4444444 0.011116804  -4.9974397 5.809649e-07 0.4226559 0.4662330
#> 6  0.5075226 0.011196015   0.6718969 5.016493e-01 0.4855788 0.5294664
#> 7  0.5655860 0.011070041   5.9246425 3.129775e-09 0.5438892 0.5872829
#> 8  0.5888945 0.011105559   8.0045093 1.199440e-15 0.5671281 0.6106610
#> 9  0.5812686 0.010982490   7.3998324 1.363565e-13 0.5597433 0.6027939
#> 10 0.5875280 0.008232239  10.6323404 2.107561e-26 0.5713931 0.6036629
#> 11 0.5261628 0.008513352   3.0731479 2.118135e-03 0.5094769 0.5428487
#> 12 0.4592973 0.008422390  -4.8326730 1.347120e-06 0.4427898 0.4758049
#> 13 0.4243392 0.008423190  -8.9824358 2.648376e-19 0.4078301 0.4408484

# marginal means with design constraints
mm(immigration, ChosenImmigrant ~ Gender + LanguageSkills + PriorEntry +
     CountryOfOrigin * ReasonForApplication, id = ~CaseID)

#>            outcome statistic                feature                    level
#> 1  ChosenImmigrant        mm                 Gender                   Female
#> 2  ChosenImmigrant        mm                 Gender                     Male
#> 3  ChosenImmigrant        mm        Language Skills           Fluent English
#> 4  ChosenImmigrant        mm        Language Skills           Broken English
#> 5  ChosenImmigrant        mm        Language Skills Tried English but Unable
#> 6  ChosenImmigrant        mm        Language Skills         Used Interpreter
#> 7  ChosenImmigrant        mm            Prior Entry                    Never
#> 8  ChosenImmigrant        mm            Prior Entry          Once as Tourist
#> 9  ChosenImmigrant        mm            Prior Entry    Many Times as Tourist
#> 10 ChosenImmigrant        mm            Prior Entry   Six Months with Family
#> 11 ChosenImmigrant        mm            Prior Entry   Once w/o Authorization
#> 12 ChosenImmigrant        mm      Country of Origin                    India
#> 13 ChosenImmigrant        mm      Country of Origin                  Germany
#> 14 ChosenImmigrant        mm      Country of Origin                   France
#> 15 ChosenImmigrant        mm      Country of Origin                   Mexico
#> 16 ChosenImmigrant        mm      Country of Origin              Philippines
#> 17 ChosenImmigrant        mm      Country of Origin                   Poland
#> 18 ChosenImmigrant        mm      Country of Origin                    China
#> 19 ChosenImmigrant        mm      Country of Origin                    Sudan
#> 20 ChosenImmigrant        mm      Country of Origin                  Somalia
#> 21 ChosenImmigrant        mm      Country of Origin                     Iraq
#> 22 ChosenImmigrant        mm Reason for Application      Reunite with Family
#> 23 ChosenImmigrant        mm Reason for Application          Seek Better Job
#> 24 ChosenImmigrant        mm Reason for Application       Escape Persecution
#>     estimate   std.error        z             p     lower     upper
#> 1  0.5117812 0.005919810 86.45230  0.000000e+00 0.5001786 0.5233838
#> 2  0.4877013 0.006048272 80.63481  0.000000e+00 0.4758469 0.4995557
#> 3  0.5875280 0.008232199 71.36950  0.000000e+00 0.5713931 0.6036628
#> 4  0.5261628 0.008513312 61.80471  0.000000e+00 0.5094770 0.5428486
#> 5  0.4592973 0.008422350 54.53316  0.000000e+00 0.4427898 0.4758048
#> 6  0.4243392 0.008423150 50.37774  0.000000e+00 0.4078302 0.4408483
#> 7  0.4860659 0.009508543 51.11886  0.000000e+00 0.4674295 0.5047023
#> 8  0.5413075 0.009443899 57.31822  0.000000e+00 0.5227978 0.5598172
#> 9  0.5427117 0.009498116 57.13889  0.000000e+00 0.5240958 0.5613277
#> 10 0.5582216 0.009328434 59.84087  0.000000e+00 0.5399382 0.5765050
#> 11 0.3730552 0.009094202 41.02121  0.000000e+00 0.3552309 0.3908795
#> 12 0.5017844 0.013358303 37.56349  0.000000e+00 0.4756026 0.5279662
#> 13 0.5425158 0.013206718 41.07878  0.000000e+00 0.5166311 0.5684005
#> 14 0.5166436 0.013159814 39.25918  0.000000e+00 0.4908508 0.5424363
#> 15 0.5124633 0.013534149 37.86447  0.000000e+00 0.4859369 0.5389898
#> 16 0.5316092 0.013374696 39.74739  0.000000e+00 0.5053953 0.5578231
#> 17 0.5267666 0.013339233 39.49002  0.000000e+00 0.5006222 0.5529110
#> 18 0.4996471 0.013282753 37.61624  0.000000e+00 0.4736134 0.5256809
#> 19 0.4809763 0.013386991 35.92863 1.091490e-282 0.4547383 0.5072143
#> 20 0.4732716 0.013329776 35.50484 4.138962e-276 0.4471457 0.4993974
#> 21 0.4086103 0.013509854 30.24535 6.005900e-201 0.3821314 0.4350891
#> 22 0.5193915 0.006459859 80.40292  0.000000e+00 0.5067304 0.5320526
#> 23 0.4759006 0.006405485 74.29580  0.000000e+00 0.4633461 0.4884552
#> 24 0.5160611 0.011467968 45.00022  0.000000e+00 0.4935843 0.5385379

# higher-order marginal means with feature interactions
immigration$language_entry <-
  interaction(immigration$LanguageSkills, immigration$PriorEntry, sep = "_")
mm(immigration, ChosenImmigrant ~ language_entry,
   id = ~CaseID)

#>            outcome statistic        feature
#> 1  ChosenImmigrant        mm language_entry
#> 2  ChosenImmigrant        mm language_entry
#> 3  ChosenImmigrant        mm language_entry
#> 4  ChosenImmigrant        mm language_entry
#> 5  ChosenImmigrant        mm language_entry
#> 6  ChosenImmigrant        mm language_entry
#> 7  ChosenImmigrant        mm language_entry
#> 8  ChosenImmigrant        mm language_entry
#> 9  ChosenImmigrant        mm language_entry
#> 10 ChosenImmigrant        mm language_entry
#> 11 ChosenImmigrant        mm language_entry
#> 12 ChosenImmigrant        mm language_entry
#> 13 ChosenImmigrant        mm language_entry
#> 14 ChosenImmigrant        mm language_entry
#> 15 ChosenImmigrant        mm language_entry
#> 16 ChosenImmigrant        mm language_entry
#> 17 ChosenImmigrant        mm language_entry
#> 18 ChosenImmigrant        mm language_entry
#> 19 ChosenImmigrant        mm language_entry
#> 20 ChosenImmigrant        mm language_entry
#>                                              level  estimate  std.error
#> 1                             Fluent English_Never 0.5963939 0.01827231
#> 2                             Broken English_Never 0.5000000 0.01925997
#> 3                   Tried English but Unable_Never 0.4211248 0.01828730
#> 4                           Used Interpreter_Never 0.4209703 0.01953176
#> 5                   Fluent English_Once as Tourist 0.6330409 0.01842944
#> 6                   Broken English_Once as Tourist 0.5700000 0.01871277
#> 7         Tried English but Unable_Once as Tourist 0.5195936 0.01903453
#> 8                 Used Interpreter_Once as Tourist 0.4458509 0.01864185
#> 9             Fluent English_Many Times as Tourist 0.6222222 0.01806926
#> 10            Broken English_Many Times as Tourist 0.5759768 0.01880069
#> 11  Tried English but Unable_Many Times as Tourist 0.5082956 0.01941641
#> 12          Used Interpreter_Many Times as Tourist 0.4579025 0.01914901
#> 13           Fluent English_Six Months with Family 0.6250000 0.01804284
#> 14           Broken English_Six Months with Family 0.6008902 0.01886381
#> 15 Tried English but Unable_Six Months with Family 0.5243902 0.01838400
#> 16         Used Interpreter_Six Months with Family 0.4843305 0.01886269
#> 17           Fluent English_Once w/o Authorization 0.4651163 0.01844877
#> 18           Broken English_Once w/o Authorization 0.3865906 0.01839320
#> 19 Tried English but Unable_Once w/o Authorization 0.3211144 0.01787934
#> 20         Used Interpreter_Once w/o Authorization 0.3151261 0.01738657
#>           z             p     lower     upper
#> 1  32.63922 1.139567e-233 0.5605808 0.6322070
#> 2  25.96058 1.381107e-148 0.4622512 0.5377488
#> 3  23.02826 2.429363e-117 0.3852824 0.4569673
#> 4  21.55312 4.949644e-103 0.3826887 0.4592518
#> 5  34.34943 1.436099e-258 0.5969199 0.6691620
#> 6  30.46047 8.704228e-204 0.5333236 0.6066764
#> 7  27.29742 4.551771e-164 0.4822866 0.5569006
#> 8  23.91667 2.054415e-126 0.4093136 0.4823883
#> 9  34.43541 7.446199e-260 0.5868071 0.6576373
#> 10 30.63594 4.067899e-206 0.5391282 0.6128255
#> 11 26.17866 4.651195e-151 0.4702402 0.5463511
#> 12 23.91259 2.265590e-126 0.4203711 0.4954339
#> 13 34.63978 6.367671e-263 0.5896367 0.6603633
#> 14 31.85413 1.153859e-222 0.5639178 0.6378626
#> 15 28.52427 5.858828e-179 0.4883583 0.5604222
#> 16 25.67664 2.132002e-145 0.4473603 0.5213007
#> 17 25.21124 3.016039e-140 0.4289574 0.5012752
#> 18 21.01812  4.478211e-98 0.3505406 0.4226406
#> 19 17.96008  4.002523e-72 0.2860715 0.3561572
#> 20 18.12468  2.035187e-73 0.2810490 0.3492031