Marginal Means

Calculate (descriptive) marginal means (MMs) from a conjoint design

mm(data, formula, id = NULL, weights = NULL, feature_order = NULL,
  feature_labels = NULL, level_order = c("ascending", "descending"),
  h0 = 0, alpha = 0.05, ...)

Arguments

data
A data frame containing variables specified in formula. All RHS variables should be factors.

formula
A formula specifying an outcome (LHS) and conjoint features (RHS) to describe. All RHS variables should be factors.

id
An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.

weights
An (optional) RHS formula specifying a variable holding survey weights.

feature_order
An (optional) character vector specifying the names of feature (RHS) variables in the order they should be encoded in the resulting data frame.

feature_labels
A named list of “fancy” feature labels to be used in output. By default, the function looks for a “label” attribute on each variable in formula and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables on the righthand side of formula and character string values; arguments passed here override variable attributes.

level_order
A character string specifying levels (within each feature) should be ordered increasing or decreasing in the final output. This is mostly only consequential for plotting via plot.cj_mm, etc.

h0
A numeric value specifying a null hypothesis value to use when generating z-statistics and p-values.

alpha
A numeric value indicating the significance level at which to calculate confidence intervals for the MMs (by default 0.95, meaning 95-percent CIs are returned).

…
Ignored.

Details

mm provides descriptive representations of conjoint data as marginal means (MMs), which represent the mean outcome across all appearances of a particular conjoint feature level, averaging across all other features. In forced choice conjoint designs, MMs by definition average 0.5 with values above 0.5 indicating features that increase profile favorability and values below 0.5 indicating features that decrease profile favorability. For continuous outcomes, MMs can take any value in the full range of the outcome.

But note that if feature levels can co-occur, such that both alternatives share a feature level, then the MMs on forced choice outcomes are bounded by the probability of co-occurrence (as a lower bound) and 1 minus that probability as an upper bound.

Plotting functionality is provided in plot.cj_mm.

See also

Examples

data(immigration)
mm(immigration, ChosenImmigrant ~ Gender + Education + LanguageSkills,
   id = ~ CaseID, h0 = 0.5)

#>            outcome statistic                feature                    level
#> 1  ChosenImmigrant        mm                 Gender                   Female
#> 2  ChosenImmigrant        mm                 Gender                     Male
#> 3  ChosenImmigrant        mm Educational Attainment                No Formal
#> 4  ChosenImmigrant        mm Educational Attainment                4th Grade
#> 5  ChosenImmigrant        mm Educational Attainment                8th Grade
#> 6  ChosenImmigrant        mm Educational Attainment              High School
#> 7  ChosenImmigrant        mm Educational Attainment         Two-Year College
#> 8  ChosenImmigrant        mm Educational Attainment           College Degree
#> 9  ChosenImmigrant        mm Educational Attainment          Graduate Degree
#> 10 ChosenImmigrant        mm        Language Skills           Fluent English
#> 11 ChosenImmigrant        mm        Language Skills           Broken English
#> 12 ChosenImmigrant        mm        Language Skills Tried English but Unable
#> 13 ChosenImmigrant        mm        Language Skills         Used Interpreter
#>     estimate   std.error           z            p     lower     upper
#> 1  0.5117812 0.005919838   1.9901230 4.657739e-02 0.5001785 0.5233839
#> 2  0.4877013 0.006048301  -2.0334110 4.201102e-02 0.4758469 0.4995558
#> 3  0.3895112 0.011003561 -10.0411853 1.004597e-23 0.3679446 0.4110778
#> 4  0.4222002 0.010994931  -7.0759699 1.484071e-12 0.4006505 0.4437499
#> 5  0.4444444 0.011116804  -4.9974397 5.809649e-07 0.4226559 0.4662330
#> 6  0.5075226 0.011196015   0.6718969 5.016493e-01 0.4855788 0.5294664
#> 7  0.5655860 0.011070041   5.9246425 3.129775e-09 0.5438892 0.5872829
#> 8  0.5888945 0.011105559   8.0045093 1.199440e-15 0.5671281 0.6106610
#> 9  0.5812686 0.010982490   7.3998324 1.363565e-13 0.5597433 0.6027939
#> 10 0.5875280 0.008232239  10.6323404 2.107561e-26 0.5713931 0.6036629
#> 11 0.5261628 0.008513352   3.0731479 2.118135e-03 0.5094769 0.5428487
#> 12 0.4592973 0.008422390  -4.8326730 1.347120e-06 0.4427898 0.4758049
#> 13 0.4243392 0.008423190  -8.9824358 2.648376e-19 0.4078301 0.4408484