Tax Preference Conjoint Experiment Dataset from Ballard-Rosa et al. (2016)

A dataset containing the results of a fully randomized conjoint survey of a representative sample of 2000 American adults who were asked to choose between alternative tax rate policies. Variables taxrate1-taxrate6 refer to tax rates for different income brackets and taxrev refers to levels of total tax revenue.

taxes

Format

A data frame (with additional “cj_df” class) with 32000 observations on the following 13 variables. Each row corresponds to a single profile presented to a respondent.

chose_plan

A numeric vector denoting whether the immigrant profile was selected (=1) or not (=0).

taxrate1

An experimental factor with levels “<10k: 0%”, “<10k: 5%”, “<10k: 15%”, “<10k: 25%”.

taxrate2

An experimental factor with levels “10-35k: 5%”, “10-35k: 15%”, “10-35k: 25%”, “10-35k: 35%”.

taxrate3

An experimental factor with levels “35-85k: 5%”, “35-85k: 15%”, “35-85k: 25%”, “35-85k: 35%”.

taxrate4

An experimental factor with levels “85-175k: 5%”, “85-175k: 15%”, “85-175k: 25%”, “85-175k: 35%”.

taxrate5

An experimental factor with levels “175-375k: 5%”, “175-375k: 15%”, “175-375k: 25%”, “175-375k: 35%”, “175-375k: 45%”.

taxrate6

An experimental factor with levels “>375k: 5%”, “>375k: 15%”, “>375k: 25%”, “>375k: 35%”, “>375k: 45%”, “>375k: 55%”.

taxrev

An experimental factor with levels “<75%”, “75-95%”, “95-105%”, “105-125%”, “>125%”.

inequality_aversion

A covariate specifying whether respondent is inequality averse (=1) or not (=0).

taxes_harm_economy

A covariate specifying whether respondent believes taxes harm the economy (=1) or not (=0).

partyid

A factor specifying the respondent's party identification; one of “Independent”, “Democrat”, “Republican”.

ID

A numeric vector indicating the respondent to which the profile corresponds.

weight

A numeric vector containing survey weights.

Source

Ballard-Rosa, Cameron, Lucy Martin, and Kenneth Scheve. 2016. “The Structure of American Income Tax Policy Preferences.” The Journal of Politics 79(1): 1-16. http://doi.org/10.1086/687324

See also

Examples

data("taxes")
f1 <- chose_plan ~ taxrate1 + taxrate2 + taxrate3 +
        taxrate4 + taxrate5 + taxrate6 + taxrev
cj(taxes, f1, id = ~ ID, weights = ~     weight)

#>       outcome statistic                        feature         level
#> 1  chose_plan      amce          Tax rate for <$10,000      <10k: 0%
#> 2  chose_plan      amce          Tax rate for <$10,000      <10k: 5%
#> 3  chose_plan      amce          Tax rate for <$10,000     <10k: 15%
#> 4  chose_plan      amce          Tax rate for <$10,000     <10k: 25%
#> 5  chose_plan      amce   Tax rate for $10,000-$35,000    10-35k: 5%
#> 6  chose_plan      amce   Tax rate for $10,000-$35,000   10-35k: 15%
#> 7  chose_plan      amce   Tax rate for $10,000-$35,000   10-35k: 25%
#> 8  chose_plan      amce   Tax rate for $10,000-$35,000   10-35k: 35%
#> 9  chose_plan      amce   Tax rate for $25,000-$85,000    35-85k: 5%
#> 10 chose_plan      amce   Tax rate for $25,000-$85,000   35-85k: 15%
#> 11 chose_plan      amce   Tax rate for $25,000-$85,000   35-85k: 25%
#> 12 chose_plan      amce   Tax rate for $25,000-$85,000   35-85k: 35%
#> 13 chose_plan      amce  Tax rate for $85,000-$175,000   85-175k: 5%
#> 14 chose_plan      amce  Tax rate for $85,000-$175,000  85-175k: 15%
#> 15 chose_plan      amce  Tax rate for $85,000-$175,000  85-175k: 25%
#> 16 chose_plan      amce  Tax rate for $85,000-$175,000  85-175k: 35%
#> 17 chose_plan      amce Tax rate for $175,000-$375,000  175-375k: 5%
#> 18 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 15%
#> 19 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 25%
#> 20 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 35%
#> 21 chose_plan      amce Tax rate for $175,000-$375,000 175-375k: 45%
#> 22 chose_plan      amce         Tax rate for >$375,000     >375k: 5%
#> 23 chose_plan      amce         Tax rate for >$375,000    >375k: 15%
#> 24 chose_plan      amce         Tax rate for >$375,000    >375k: 25%
#> 25 chose_plan      amce         Tax rate for >$375,000    >375k: 35%
#> 26 chose_plan      amce         Tax rate for >$375,000    >375k: 45%
#> 27 chose_plan      amce         Tax rate for >$375,000    >375k: 55%
#> 28 chose_plan      amce                    Tax revenue          <75%
#> 29 chose_plan      amce                    Tax revenue        75-95%
#> 30 chose_plan      amce                    Tax revenue       95-105%
#> 31 chose_plan      amce                    Tax revenue      105-125%
#> 32 chose_plan      amce                    Tax revenue         >125%
#>         estimate   std.error            z             p         lower
#> 1   0.0000000000          NA           NA            NA            NA
#> 2  -0.0213657909 0.005603362  -3.81303088  1.372730e-04 -0.0323481778
#> 3  -0.1085089681 0.006796459 -15.96551498  2.222001e-57 -0.1218297830
#> 4  -0.2417205854 0.008531343 -28.33324048 1.346689e-176 -0.2584417099
#> 5   0.0000000000          NA           NA            NA            NA
#> 6  -0.0216654811 0.007365031  -2.94166851  3.264492e-03 -0.0361006774
#> 7  -0.1028048639 0.010913360  -9.42009287  4.506892e-21 -0.1241946563
#> 8  -0.2194266402 0.013745146 -15.96393661  2.278927e-57 -0.2463666314
#> 9   0.0000000000          NA           NA            NA            NA
#> 10  0.0001812976 0.005518348   0.03285361  9.737913e-01 -0.0106344656
#> 11 -0.0650669525 0.007085019  -9.18373729  4.163836e-20 -0.0789533344
#> 12 -0.1260348655 0.008661017 -14.55197102  5.674421e-48 -0.1430101465
#> 13  0.0000000000          NA           NA            NA            NA
#> 14  0.0109438886 0.005308487   2.06158349  3.924741e-02  0.0005394458
#> 15 -0.0155419502 0.005542475  -2.80415325  5.044892e-03 -0.0264050022
#> 16 -0.0126122794 0.005761680  -2.18899344  2.859732e-02 -0.0239049645
#> 17  0.0000000000          NA           NA            NA            NA
#> 18  0.0332708646 0.005571528   5.97158679  2.349570e-09  0.0223508699
#> 19  0.0425251595 0.006516894   6.52537219  6.783288e-11  0.0297522818
#> 20  0.0580764422 0.005944731   9.76939761  1.523552e-22  0.0464249833
#> 21  0.0327895914 0.006830710   4.80031956  1.584127e-06  0.0194016456
#> 22  0.0000000000          NA           NA            NA            NA
#> 23  0.0558225384 0.006793283   8.21731380  2.081120e-16  0.0425079483
#> 24  0.0896621795 0.007361044  12.18063394  3.942058e-34  0.0752347989
#> 25  0.1139343820 0.007379616  15.43906689  8.938405e-54  0.0994706008
#> 26  0.1099558977 0.008312893  13.22715226  6.116108e-40  0.0936629273
#> 27  0.0838214633 0.008780819   9.54597283  1.348361e-21  0.0666113735
#> 28  0.0000000000          NA           NA            NA            NA
#> 29  0.0297942004 0.008374721   3.55763502  3.742088e-04  0.0133800490
#> 30  0.0692814207 0.011332045   6.11376152  9.730969e-10  0.0470710206
#> 31  0.0911924955 0.014408949   6.32887902  2.469486e-10  0.0629514743
#> 32  0.1354902736 0.018772915   7.21732757  5.301916e-13  0.0986960372
#>           upper
#> 1            NA
#> 2  -0.010383404
#> 3  -0.095188153
#> 4  -0.224999461
#> 5            NA
#> 6  -0.007230285
#> 7  -0.081415072
#> 8  -0.192486649
#> 9            NA
#> 10  0.010997061
#> 11 -0.051180571
#> 12 -0.109059585
#> 13           NA
#> 14  0.021348331
#> 15 -0.004678898
#> 16 -0.001319594
#> 17           NA
#> 18  0.044190859
#> 19  0.055298037
#> 20  0.069727901
#> 21  0.046177537
#> 22           NA
#> 23  0.069137128
#> 24  0.104089560
#> 25  0.128398163
#> 26  0.126248868
#> 27  0.101031553
#> 28           NA
#> 29  0.046208352
#> 30  0.091491821
#> 31  0.119433517
#> 32  0.172284510