Tax Preference Conjoint Experiment Dataset from Ballard-Rosa et al. (2016)

A dataset containing the results of a conjoint survey of a representative sample of 2000 American adults who were asked to choose between alternative tax rate policies. Variables taxrate1-taxrate6 refer to tax rates for different income brackets and taxrev refers to levels of total tax revenue.

taxes

Format

A data frame with 32000 observations on the following 13 variables. Each row corresponds to a single profile presented to a respondent.

chose_plan

A numeric vector denoting whether the immigrant profile was selected (=1) or not (=0).

taxrate1

An experimental factor with levels “<10k: 0%”, dQuote<10k: 5%, dQuote<10k: 15%, dQuote<10k: 25%.

taxrate2

An experimental factor with levels “10-35k: 5%”, dQuote10-35k: 15%, dQuote10-35k: 25%, dQuote10-35k: 35%.

taxrate3

An experimental factor with levels “35-85k: 5%”, dQuote35-85k: 15%, dQuote35-85k: 25%, dQuote35-85k: 35%.

taxrate4

An experimental factor with levels “85-175k: 5%”, dQuote85-175k: 15%, dQuote85-175k: 25%, dQuote85-175k: 35%.

taxrate5

An experimental factor with levels “175-375k: 5%”, dQuote175-375k: 15%, dQuote175-375k: 25%, dQuote175-375k: 35%, dQuote175-375k: 45%.

taxrate6

An experimental factor with levels “>375k: 5%”, “>375k: 15%”, “>375k: 25%”, “>375k: 35%”, “>375k: 45%”, “>375k: 55%”.

taxrev

An experimental factor with levels “<75%”, “75-95%”, “95-105%”, “105-125%”, “>125%”.

inequality_aversion

A covariate specifying whether respondent is inequality averse (=1) or not (=0).

taxes_harm_economy

A covariate specifying whether respondent believes taxes harm the economy (=1) or not (=0).

partyid

A factor specifying the respondent's party identification; one of “Independent”, “Democrat”, “Republican”.

ID

A numeric vector indicating the respondent to which the profile corresponds.

weight

A numeric vector containing survey weights.

Source

Ballard-Rosa, Cameron, Lucy Martin, and Kenneth Scheve. 2016. “The Structure of American Income Tax Policy Preferences.” The Journal of Politics 79(1): 1-16. http://doi.org/10.1086/687324

See also

Examples

# NOT RUN {
data("taxes")
f1 <- chose_plan ~ taxrate1 + taxrate2 + taxrate3 +
        taxrate4 + taxrate5 + taxrate6 + taxrev
cj(taxes, f1, id = ~ ID, weights = weight)
# }